requirements.txt
flask
opencv-python
pyserial
pyyaml
flask-socketio
flask-cors
flask-restful